Grid with capability of Shrinking, not done for second version

+++ src/org/academiadecodigo/haltistas/AwesomeGame/player/PlayerPosition.java
@@ -42,6 +42,10 @@ public class PlayerPosition {
     }
 
     public void deleteAp() {
+        if (apple == null) {
+            return;
+        }
+
         apple.delete();
         apple = null;
